extract parseoptions
